ndvi-engine/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
earthengine-api==0.1.392
numpy==1.26.4
python-dotenv==1.0.1
requests==2.31.0
